requirements.txt
anthropic>=0.40.0
python-dotenv>=1.0.0
Pillow>=10.0.0
click>=8.1.0
Django>=5.0
django-cors-headers>=4.3.0
